requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
pillow>=8.3.0
opencv-python>=4.5.0
matplotlib>=3.4.0
tensorboard>=2.8.0
tqdm>=4.62.0
einops>=0.4.0
kornia>=0.6.0
albumentations>=1.0.0
rasterio>=1.2.0
# GDAL should be installed separately using conda or pre-built wheels
# For Windows: Download GDAL wheel from https://www.lfd.uci.edu/~gohlke/pythonlibs/#gdal
scikit-image>=0.18.0
lpips>=0.1.4
wandb>=0.12.0
pyyaml>=6.0
pandas>=1.3.0
seaborn>=0.11.0
Pillow>=8.0.0
rasterio>=1.3.0
albumentations>=1.3.0
tqdm>=4.65.0
wandb>=0.15.0
gdown>=4.7.0
scikit-image>=0.19.0
lpips>=0.1.4
pytorch-fid>=0.3.0 